refactor: MovieHeader SearchInput 파일 위치 변경
