+++ src/index.js
@@ -1,4 +1,4 @@
 import "../templates/reset.css";
 import "../templates/common.css";
 
-import "./component/Header";
+import "./component/head/MovieHeader";
